Move profile stage components under /stages
