+++ frontend/src/components/stages/profile_participant_editor.scss
