+++ frontend/src/components/stages/profile_participant_view.scss
